refactor : SCSS 파일 styles 폴더 아래로 이동
