+++ client/src/utils/style/FloatingReqCreateButton.scss
@@ -1,5 +1,5 @@
-@import "../utils/style/color";
-@import "../utils/style/size";
+@import "./color";
+@import "./size";
 
 .text-box {
   display: flex;
