+++ client/src/utils/style/MainPage.scss
