+++ client/src/utils/style/RequestListPagenation.scss
@@ -1,5 +1,5 @@
-@import "../utils/style/color";
-@import "../utils/style/size";
+@import "./color";
+@import "./size";
 
 .pagenation {
   display: flex;
