+++ client/src/utils/style/RequestProject.scss
@@ -1,5 +1,5 @@
-@import "../utils/style/color";
-@import "../utils/style/size";
+@import "./color";
+@import "./size";
 
 .requestBox {
   display: flex;
