+++ client/src/utils/style/contributeList.scss
@@ -1,5 +1,5 @@
-@import "../utils/style/color";
-@import "../utils/style/size";
+@import "./color";
+@import "./size";
 
 .contributeList-box {
   display: flex;
