+++ client/src/utils/style/create.scss
@@ -1,5 +1,5 @@
-@import "../utils/style/color";
-@import "../utils/style/size";
+@import "./color";
+@import "./size";
 
 .box {
   display: flex;
