--- client/src/view/NavBar.scss
@@ -1,29 +1,0 @@
-header {
-    height: 100px;
-    display: flex;
-    padding: 20px 160px;
-
-    .logo {
-        display: inline-block;
-        color: #ff8900;
-    }
-
-    a {
-        text-decoration: none;
-        color: black;
-        cursor: pointer;
-    }
-
-    nav {
-        margin-left: auto;
-
-        ul {
-            display: flex;
-            list-style: none;
-        }
-
-        li {
-            padding: 0 15px;
-        }
-    }
-}
