+++ client/src/view/RequestListView.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { User } from "../types/userType";
 import { RequestProjectType } from "../types/requestProjectType";
 import { RequestProjectView } from "./RequestProjectView";
-import "./RequestList.scss";
+import "../utils/style/RequestList.scss";
 
 type RequestListProps = {
   requestList: Array<User & RequestProjectType>;
